backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
pyjwt==2.9.0
bcrypt==4.2.0
